[#42] refactor: 경로 변경
